KL-149/feat: add SubcategoryService interface
